+++ src/com/illinois/rts/visualizer/TaskContainer.java
@@ -82,7 +82,7 @@ public class TaskContainer {
         }
         maxId++;
 
-        addTask(maxId, "Task" + maxId, Task.TASK_TYPE_APP, 10*ProgConfig.TIMESTAMP_MS_TO_UNIT_MULTIPLIER, 10*ProgConfig.TIMESTAMP_MS_TO_UNIT_MULTIPLIER, 1*ProgConfig.TIMESTAMP_MS_TO_UNIT_MULTIPLIER, 0);
+        addTask(maxId, "Task" + maxId, Task.TASK_TYPE_APP, (int)(10*ProgConfig.TIMESTAMP_MS_TO_UNIT_MULTIPLIER), (int)(10*ProgConfig.TIMESTAMP_MS_TO_UNIT_MULTIPLIER),(int) (1*ProgConfig.TIMESTAMP_MS_TO_UNIT_MULTIPLIER), 0);
         return getTaskById(maxId);
     }
